cgi-bin/requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
arabic-reshaper==2.1.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
colorama==0.4.4
future==0.18.2
geoip2==4.4.0
google-api-core==2.2.0
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
h11==0.12.0
holehe==1.60.1
html5lib==1.1
httpcore==0.13.7
httplib2==0.20.1
httpx==0.20.0
idna==3.3
install==1.3.4
maxminddb==2.2.0
multidict==5.2.0
outcome==1.1.0
Pillow==8.4.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
reportlab==3.6.2
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
shodan==1.25.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
termcolor==1.1.0
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
virustotal-api==1.1.11
webencodings==0.5.1
xhtml2pdf==0.2.5
XlsxWriter==3.0.1
yarl==1.7.0
